requirements.txt
# Docling
docling

streamlit
pandas
python-dotenv

# Pydantic AI stack
pydantic
pydantic-ai

# HF ecosystem (important: keep older compatible versions)
transformers
huggingface-hub
tokenizers


# Optional but often required by docling
torch
numpy